remove some more unused code
